Improve code consistency and readability
Various changes has been made to different files in order to enhance the overall consistency and readability of the code. The changes involved adhering to style guide rules for indentation, line lengths, and quotations. In addition, a section for Code Quality Checks was added to Readme for better guidance. The server's URL was also updated in client's package.json to reflect the switch to a Docker environment and a typing clarification was done on response.py.